requirements.txt
bcrypt==4.0.1
black==23.3.0
cryptography==40.0.2
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
itsdangerous==2.0.0
psycopg2-binary==2.9.6
pycryptodome==3.17
pydantic==1.10.7
python-dotenv
redis==4.5.5
requests==2.29.0
SQLAlchemy==2.0.10
